change stucture database

+++ src/main/java/com/fresh/coding/sooatelapi/dtos/MostMenu.java
@@ -14,10 +14,13 @@ public  class MostMenu {
     private Long id;
     private String name;
     private String description;
-    private Double price;
+    private Long price;
     @JsonIgnore
     private Long categoryId;
     @JsonIgnore
     private String categoryName;
-    private Double quantityMenuByName;
+    private Long quantityMenuByName;
 }
+
+
+
